+++ src/View/extra-block.js
@@ -1,0 +1,6 @@
+export const createExtraBlockTemplate = () => (
+  `<section class="films-list--extra">
+    <h2 class="films-list__title">Top rated</h2>
+    <div class="films-list__container"></div>
+  </section>`
+);
