+++ src/View/footer-stats.js
@@ -1,0 +1,5 @@
+export const createFooterStatsTemplate = () => (
+  `<section class="footer__statistics">
+    <p>130 291 movies inside</p>
+  </section>`
+);
