+++ src/View/load-button.js
@@ -1,0 +1,3 @@
+export const createLoadButtonTemplate = () => (
+  `<button class="films-list__show-more">Show more</button>`
+);
