+++ src/util.js
@@ -1,0 +1,34 @@
+const SITE_BODY = document.querySelector(`body`);
+
+const IterationCount = Object.freeze({
+  DEFAULT: 1,
+  CARD: 5,
+  EXTRA: 2,
+  COMMENT: 4
+});
+
+const InsertPlace = Object.freeze({
+  BEFOREEND: `beforeend`,
+  AFTERBEGIN: `afterbegin`,
+  AFTEREND: `afterend`
+});
+
+const SiteElements = Object.freeze({
+  BODY: SITE_BODY,
+  HEADER: SITE_BODY.querySelector(`.header`),
+  MAIN: SITE_BODY.querySelector(`main`),
+  FOOTER: SITE_BODY.querySelector(`.footer`)
+});
+
+const render = ({
+  container,
+  template,
+  place = InsertPlace.BEFOREEND,
+  iteration = IterationCount.DEFAULT
+}) => {
+  for (let i = 0; i < iteration; i++) {
+    container.insertAdjacentHTML(place, template);
+  }
+};
+
+export {IterationCount, InsertPlace, SiteElements, render};
